remove json2csv function from utils.js
